+++ void_logging/rlgym_learn/__init__.py
@@ -1,5 +1,6 @@
 from .reward_logger import RewardLogger, reward_logger_serde
 from .reward_shared_info_provider import RewardSharedInfoProvider
 from .metric_logger import CustomMetricLogger, custom_metrics_serde
+from .multi_logger import MultiLogger
 
-__all__ = ["RewardLogger", "RewardSharedInfoProvider", "CustomMetricLogger", "custom_metrics_serde", "reward_logger_serde"]
+__all__ = ["RewardLogger", "RewardSharedInfoProvider", "CustomMetricLogger", "custom_metrics_serde", "reward_logger_serde", "MultiLogger"]